Первый комит: приступаю. Коды добавил. Переписал. Добавил геттеры и сеттеры. Правкт по замечянию наставника.
