+++ src/ru/skypro/Car.java
@@ -1,25 +1,9 @@
 package ru.skypro;
 
-public class Car {
+public class Car extends WheelableTransport {
 
-    private String modelName;
-    private int wheelsCount;
-
-
-    public String getModelName() {
-        return modelName;
-    }
-
-    public void setModelName(String modelName) {
-        this.modelName = modelName;
-    }
-
-    public int getWheelsCount() {
-        return wheelsCount;
-    }
-
-    public void setWheelsCount(int wheelsCount) {
-        this.wheelsCount = wheelsCount;
+    public Car(String modelName, int wheelsCount) {
+        super(modelName, wheelsCount);
     }
 
     public void updateTyre() {
