+++ src/ru/skypro/WheelableTransport.java
@@ -1,0 +1,27 @@
+package ru.skypro;
+
+public abstract class WheelableTransport {
+    private String modelName;
+    private int wheelsCount;
+
+    public WheelableTransport(String modelName, int wheelsCount) {
+        this.modelName = modelName;
+        this.wheelsCount = wheelsCount;
+    }
+
+    public String getModelName() {
+        return modelName;
+    }
+
+    public void setModelName(String modelName) {
+        this.modelName = modelName;
+    }
+
+    public int getWheelsCount() {
+        return wheelsCount;
+    }
+
+    public void setWheelsCount(int wheelsCount) {
+        this.wheelsCount = wheelsCount;
+    }
+}
